correction update et ds

+++ updates/create_waka_mails_table_u170.php
@@ -8,25 +8,29 @@ class CreateWakaMailsTableU170 extends Migration
 {
     public function up()
     {
-        Schema::table('waka_mailer_waka_mails', function (Blueprint $table) {
-            $table->dropColumn('model_functions');
-            $table->dropColumn('is_scope');
-            $table->dropColumn('scopes');
-            $table->dropColumn('has_asks');
-            $table->dropColumn('asks');
-            $table->dropColumn('images');
-        });
+        if (Schema::hasColumn('waka_mailer_waka_mails', 'model_functions')){
+            Schema::table('waka_mailer_waka_mails', function (Blueprint $table) {
+                $table->dropColumn('model_functions');
+                $table->dropColumn('is_scope');
+                $table->dropColumn('scopes');
+                $table->dropColumn('has_asks');
+                $table->dropColumn('asks');
+                $table->dropColumn('images');
+            });
+        }
     }
 
     public function down()
     {
-        Schema::table('waka_mailer_waka_mails', function (Blueprint $table) {
-            $table->text('model_functions')->nullable();
-            $table->boolean('is_scope')->nullable()->default(false);
-            $table->text('scopes')->nullable();
-            $table->boolean('has_asks')->nullable();
-            $table->text('asks')->nullable();
-            $table->text('images')->nullable();
-        });
+        if (!Schema::hasColumn('waka_mailer_waka_mails', 'model_functions')){
+            Schema::table('waka_mailer_waka_mails', function (Blueprint $table) {
+                $table->text('model_functions')->nullable();
+                $table->boolean('is_scope')->nullable()->default(false);
+                $table->text('scopes')->nullable();
+                $table->boolean('has_asks')->nullable();
+                $table->text('asks')->nullable();
+                $table->text('images')->nullable();
+            });
+        }
     }
 }
